requirements.txt
streamlit
google-generativeai
streamlit-agraph
python-docx